+++ include/packet_queue.hpp
@@ -10,9 +10,9 @@ class PacketQueue {
         void add_delay(int delay);
 
         int get_lost_packet_count() const;
-        double get_quant() const;
-        double get_deficit() const;
-        void set_deficit(double deficit);
+        int get_quant() const;
+        int get_deficit() const;
+        void set_deficit(int deficit);
 
         Packet front() const;
         int size() const;
@@ -22,8 +22,8 @@ class PacketQueue {
         void print();
 
     private:
-        double quant; // Время выделяемое очереди 
-        double deficit; // Накопленный дефицит времени
+        int quant; // Выделяемое очереди время в ресурсных блоках (RB)
+        int deficit; // Накопленный дефицит в ресурсных блоках (RB)
         size_t limit; // Размер очереди
         int lost_packet_count = 0; // Число потерянных пакетов
         std::priority_queue<Packet, std::vector<Packet>, PacketGreater> packet_queue; // Очередь из обслуживаемых пакетов
