+++ src/packet.cpp
@@ -2,7 +2,7 @@
 
 Packet::Packet(int size = 0) : size(size), scheduled_at(0.0), retry(false) {}
 
-double Packet::get_size() const { return size; }
+int Packet::get_size() const { return size; }
 bool Packet::get_retry() const { return retry; }
 void Packet::set_retry() { this->retry = true; }
